# api/openapi.yaml
openapi: 3.1.0
info:
  title: MailRules API
  version: 0.9.0
  description: |
    JSON REST API of the MailRules daemon, plus one Server-Sent Events stream. This file is
    the contract the web UI's TypeScript client is generated from. A test fails if the daemon
    serves an /api path that is not listed here, or the reverse.

    **Conventions**

    - JSON bodies. Unknown fields in a request are rejected with 400 `invalid_json`.
    - Timestamps are unix seconds. Ids are integers. An absent time or reference is `null`.
    - Errors are `{"error": {"code", "message", "path"}}`. `code` is stable and machine-readable;
      `message` is a sentence safe to show; `path` names the field when there is one
      (for a rule: `conditions.all[0].op`).
    - Lists that can grow use cursor pagination: `?cursor=&limit=` (limit 1 to 100, default 50).
      The response carries `next_cursor`; pass it back as `cursor`, and `null` means the end.
    - Every route except the four under `/api/auth`, `/healthz`, `/readyz` and `/metrics` needs
      the `mailrules_session` cookie, and answers 401 without one: `setup_required` before the
      admin account exists, `unauthenticated` after.
    - Every response hands out a `mailrules_csrf` cookie, and every non-GET request must send
      its value back in the `X-CSRF-Token` header, or it is answered 403 `csrf_failed`.
      Call `GET /api/auth/me` first: it hands out the cookie and says which screen to show.
...
servers:
...
security:
  - session: []
...
paths:
  /api/auth/setup:
    post:
      tags: [Auth]
      security: []
      operationId: authSetup
      summary: Create the admin account (first run only) and sign in
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Credentials' }
      responses:
        '201':
          description: Account created; session cookie set
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Session' }
        '400': { $ref: '#/components/responses/Invalid' } # invalid_input (path: email | password), invalid_json
        '403': { $ref: '#/components/responses/CsrfFailed' }
        '409': { $ref: '#/components/responses/Error' } # already_set_up
  /api/auth/login:
    post:
      tags: [Auth]
      security: []
      operationId: authLogin
      summary: Sign in; replaces any session the browser already holds
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Credentials' }
      responses:
        '200':
          description: Signed in; session cookie set
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Session' }
        '400': { $ref: '#/components/responses/Invalid' } # invalid_json
        '401': { $ref: '#/components/responses/Error' } # invalid_credentials
        '403': { $ref: '#/components/responses/CsrfFailed' }
        '429':
          description: rate_limited; five failed sign-ins per minute per IP
          headers:
            Retry-After:
              schema: { type: integer }
              description: Seconds until sign-in is allowed again
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorBody' }
  /api/auth/logout:
    post:
      tags: [Auth]
      security: []
      operationId: authLogout
      summary: End the session; safe to call when not signed in
      responses:
        '204': { description: Signed out }
        '403': { $ref: '#/components/responses/CsrfFailed' }
  /api/auth/me:
    get:
      tags: [Auth]
      security: []
      operationId: authMe
      summary: Who is signed in
      responses:
        '200':
          description: Signed in
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Session' }
        '401': { $ref: '#/components/responses/Error' } # setup_required (show first-run setup) | unauthenticated (show sign-in)
  /api/presets:
    get:
      tags: [Accounts]
      operationId: listPresets
      summary: Provider presets for the connect wizard
      responses:
        '200':
          description: Every preset, generic last
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/Preset' } }
        '401': { $ref: '#/components/responses/Unauthenticated' }
  /api/accounts/test:
    post:
      tags: [Accounts]
      operationId: testAccount
      summary: Log in with credentials that are not stored, and list the folders
      description: Changes nothing, in the database or the mailbox. The password is used for this login only.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AccountInput' }
      responses:
        '200':
          description: The server accepted the login
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AccountTestResult' }
        '400': { $ref: '#/components/responses/Invalid' } # invalid_input (path: preset | username | password | host | port | tls_mode), invalid_json
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/CsrfFailed' }
        '422': { $ref: '#/components/responses/ConnectionFailed' }
  /api/accounts:
    get:
      tags: [Accounts]
      operationId: listAccounts
      summary: Every connected mailbox, oldest first
      responses:
        '200':
          description: Accounts
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/Account' } }
        '401': { $ref: '#/components/responses/Unauthenticated' }
    post:
      tags: [Accounts]
      operationId: createAccount
      summary: Connect a mailbox and start watching it
      description: |
        The login is checked against the server before anything is stored. Only mail that
        arrives from now on is sorted; existing mail is Cleanup's job. The new account's
        `status` is `new` until its supervisor reports over `account.status`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AccountInput' }
      responses:
        '201':
          description: Stored and being watched
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AccountEnvelope' }
        '400': { $ref: '#/components/responses/Invalid' } # as for /api/accounts/test
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/CsrfFailed' }
        '409': { $ref: '#/components/responses/Conflict' } # account_exists (path: username)
        '422': { $ref: '#/components/responses/ConnectionFailed' }
  /api/accounts/{id}:
    parameters:
      - { $ref: '#/components/parameters/Id' }
    get:
      tags: [Accounts]
      operationId: getAccount
      summary: One account
      responses:
        '200':
          description: The account
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AccountEnvelope' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Accounts]
      operationId: updateAccount
      summary: Rename, change the watched folder or the app password, pause or resume
      description: Anything but the label restarts the account's connection, so the change is in force when the response arrives.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AccountPatch' }
      responses:
        '200':
          description: The account as it is now
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AccountEnvelope' }
        '400': { $ref: '#/components/responses/Invalid' } # invalid_input (path: label | watch_folder | password | paused), invalid_json
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/CsrfFailed' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Accounts]
      operationId: deleteAccount
      summary: Disconnect a mailbox and wipe everything stored about it
      description: Stops watching, then deletes the account with its password, folders, contacts, messages, decisions, actions and the rules scoped to it. Nothing in the mailbox is touched.
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/CsrfFailed' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/accounts/{id}/reconnect:
    parameters:
      - { $ref: '#/components/parameters/Id' }
    post:
      tags: [Accounts]
      operationId: reconnectAccount
      summary: Drop the connection and connect again
      description: Also how an account that stopped on `auth_failed` or `error` is started again. The outcome arrives over `account.status`.
      responses:
        '200':
          description: Reconnecting; the body is the account as stored at this moment
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AccountEnvelope' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/CsrfFailed' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' } # account_paused
  /api/accounts/{id}/folders:
    parameters:
      - { $ref: '#/components/parameters/Id' }
    get:
      tags: [Accounts]
      operationId: listAccountFolders
      summary: The account's folders, as last discovered on the server
      responses:
        '200':
          description: Folders by name
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/Folder' } }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/rules:
    get:
      tags: [Rules]
      operationId: listRules
      summary: Every rule in priority order, with this week's hit count and last match
      responses:
        '200':
          description: Rules, disabled ones included
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RuleList' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
  /api/rules/compose:
    post:
      tags: [Rules]
      operationId: composeRules
      summary: Turn free text into validated draft rules
      description: |
        Nothing is saved. Each draft comes back validated: what is wrong with one is in its
        `errors`, and the request still answers 200. The drafts without errors are then
        tested together, in the order given, on the last 200 messages of the account's
        watched folder (model calls recorded with purpose `test`; the composer's own call
        with purpose `compose`), and each gets `match_count` and up to 5 `samples`. With no
        account connected nothing is tested. Needs a generative model: 409
        `no_composer_model` until an Anthropic key is set in Settings.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ComposeRequest' }
      responses:
        '200':
          description: Drafts to review
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ComposeResult' }
        '400': { $ref: '#/components/responses/Invalid' } # invalid_input (path: text | account_id), invalid_json
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/CsrfFailed' }
        '409': { $ref: '#/components/responses/Conflict' } # no_composer_model
        '502': { $ref: '#/components/responses/Upstream' } # model_error
  /api/rules/batch:
    post:
      tags: [Rules]
      operationId: createRules
      summary: Save rules, in one transaction
      description: |
        How every rule is created besides import: approved composer drafts, a rule built by
        hand in the condition builder, a template. It needs no model. Every rule is
        validated before anything is done, so one bad rule saves none. `new_folders` are
        then created on the rule's account (every connected account for a rule with
        `account_id` null), and the rules inserted at the end or at `position`. Folders are
        not created while dry-run is on or the account is offline; the first live move
        creates a missing folder.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RuleBatchRequest' }
      responses:
        '201':
          description: The saved rules
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/Rule' } }
        '400': { $ref: '#/components/responses/Invalid' } # rule_invalid (path: rules[0].conditions.all[0].op | rules[0].model | rules[0].new_folders[0]), invalid_input (path: rules | rules[0] for a field of the wrong type or an unknown one | rules[0].account_id), invalid_json
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/CsrfFailed' }
        '502': { $ref: '#/components/responses/Upstream' } # mailbox_error: a folder could not be created; nothing was saved
  /api/rules/reorder:
    post:
      tags: [Rules]
      operationId: reorderRules
      summary: Set the priority order of all rules
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids]
              additionalProperties: false
              properties:
                ids:
                  type: array
                  items: { type: integer, format: int64 }
                  description: Every rule's id exactly once; the first runs first
      responses:
        '200':
          description: The rules in their new order
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RuleList' }
        '400': { $ref: '#/components/responses/Invalid' } # invalid_input (path: ids), invalid_json
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/CsrfFailed' }
  /api/rules/test:
    post:
      tags: [Rules]
      operationId: testRules
      summary: Run saved or draft rules over recent mail without acting
      description: |
        Never touches the mailbox: mail is read with BODY.PEEK and stays unread and where it
        is. Decides exactly as live processing does (sender rules, conditions, the decision
        model, each rule's own `model`). Model usage is recorded with purpose `test`. Rules
        with an intent need a decision model: 409 `no_composer_model` when none is set;
        condition-only rules are tested without one. With `limit` up to 200 the answer is
        one JSON body. Above that it is a `text/event-stream`: `progress` events carrying
        `TestProgress` (one per 25 emails), then one `done` event carrying `TestResult`, or,
        if the run fails midway, one `error` event carrying `ErrorBody` (code `test_failed`).
        A run that fails before the first email is answered as plain JSON.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TestRequest' }
      responses:
        '200':
          description: What each email would get
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TestResult' }
            text/event-stream:
              schema: { type: string, description: '`progress` events (TestProgress), then `done` (TestResult) or `error` (ErrorBody)' }
        '400': { $ref: '#/components/responses/Invalid' } # invalid_input (path: rule_ids | account_id | folder | limit | rules[0]), rule_invalid (path: rules[0]...), invalid_json
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/CsrfFailed' }
        '409': { $ref: '#/components/responses/Conflict' } # account_offline, no_composer_model
        '502': { $ref: '#/components/responses/Upstream' } # model_error, mailbox_error
  /api/rules/export:
    get:
      tags: [Rules]
      operationId: exportRules
      summary: Download every rule as a YAML file
      description: The PRD rule-file shape (`id` is the rule's name, `when` the intent, `match` the conditions, `unless` the exceptions). The database stays the source of truth.
      responses:
        '200':
          description: The rules file
          content:
            application/yaml:
              schema: { type: string }
        '401': { $ref: '#/components/responses/Unauthenticated' }
  /api/rules/import:
    post:
      tags: [Rules]
      operationId: importRules
      summary: Upload a rules YAML file
      description: |
        The request body is the file itself (at most 1 MB). Nothing is stored unless every
        rule in it is valid. A rule whose name (`id` in the file) is already in use replaces
        that rule's wording and keeps its place, account and id; the others are added after
        the existing rules, in file order. Rules the file does not name are left alone.
      requestBody:
        required: true
        content:
          application/yaml:
            schema: { type: string }
      responses:
        '200':
          description: All rules after the import, with how many were added and updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ImportResult' }
        '400': { $ref: '#/components/responses/Invalid' } # rule_invalid: message lists every problem, path is the first one's
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/CsrfFailed' }
        '413': { $ref: '#/components/responses/TooLarge' }
  /api/rules/{id}:
    parameters:
      - { $ref: '#/components/parameters/Id' }
    get:
      tags: [Rules]
      operationId: getRule
      summary: One rule
      responses:
        '200':
          description: The rule
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RuleEnvelope' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Rules]
      operationId: updateRule
      summary: Edit a rule; fields left out stay as they are
      description: The result is validated as a whole and the rule's `version` goes up by one. Priority is changed with `/api/rules/reorder`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RulePatch' }
      responses:
        '200':
          description: The rule as saved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RuleEnvelope' }
        '400': { $ref: '#/components/responses/Invalid' } # rule_invalid (path e.g. conditions.all[0].op, or model), invalid_input (path: a field of the wrong type, or account_id), invalid_json
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/CsrfFailed' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Rules]
      operationId: deleteRule
      summary: Delete a rule
      description: Its past decisions stay in the activity feed under the name it had (`rule_id` becomes null there). Sender rules that route to it go with it.
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/CsrfFailed' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/rules/{id}/compose:
    parameters:
      - { $ref: '#/components/parameters/Id' }
    post:
      tags: [Rules]
      operationId: recomposeRule
      summary: Re-optimize one rule from its original wording plus new text
      description: Returns a single draft to replace the rule, validated and tested like a composed one. Nothing is saved; save it with PATCH.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              additionalProperties: false
              properties:
                text: { type: string, maxLength: 4000 }
      responses:
        '200':
          description: The replacement draft
          content:
            application/json:
              schema:
                type: object
                required: [rule]
                properties:
                  rule: { $ref: '#/components/schemas/RuleDraft' }
        '400': { $ref: '#/components/responses/Invalid' } # invalid_input (path: text), invalid_json
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/CsrfFailed' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' } # no_composer_model
        '502': { $ref: '#/components/responses/Upstream' } # model_error
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: mailrules_session
      description: Set by `/api/auth/setup` and `/api/auth/login`. Non-GET requests also need the `X-CSRF-Token` header.
...
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorBody' }
...
  schemas:
    Credentials:
      type: object
      required: [email, password]
      additionalProperties: false
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 12, description: Minimum applies to setup only }
...
    Session:
      type: object
      required: [user]
      properties:
        user: { $ref: '#/components/schemas/User' }
...